Se crea ProvinciaNegocio, se valida CP y se obtiene de base las provincias para el DropDownList

+++ TP-Cuatrimestral-Equipo-15B/domino/Provincia.cs
@@ -8,7 +8,7 @@ namespace dominio
 {
     public class Provincia
     {
-        public int Id { set; get; }
+        public byte Id { set; get; }
         public string Nombre { set; get; }
         public string Codigo31662 {  set; get; }
     }
